Change `defflow` to also define the actual flow
The naming and what it does have a big mismatch IMO. In the current
implementation `defflow` actually defines a test that runs the flow, and the
flow is never actually defined.

This proposes also defining the actual flow in a variable named `name` and have
the test name reflect what it actually is: running the flow defined by `name`.

But this is a breaking change. I couldn't think of a way of fixing this another
way though.